+++ CHANGELOG.md
@@ -16,9 +16,13 @@ the code was deployed.
 - Particle fields to the Locations table in the DB (CU-2chw9e3).
 - AWS Dockerfile.
 
-###
+### Changed
 
 - DB setup script.
+- Broke up the Client and Location `is_active` fields into the component parts: `is_displayed`, `is_sending_alerts`, and `is_sending_vitals` (CU-860ptt5rp).
+- Only display Clients and Locations in the dashboard if their `is_displayed` is true (CU-860ptt5rp).
+- Only send vitals messages if the relevant `is_sending_vitals` is true (CU-860ptt5rp).
+- Only send Stillness/Duration alerts if the relevant `is_sending_vitals` is true (CU-860ptt5rp).
 
 ## [9.1.0] - 2023-03-13